pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0.0",
    "asyncpraw>=7.8.1",
    "praw>=7.8.1",
    "python-dotenv>=1.0.1",
    "twilio>=9.4.5",
]
